wj-product/src/main/java/xin/wangjue/domain/entity/Area.java
package xin.wangjue.domain.entity;


import java.sql.Date;
import java.sql.Timestamp;

public class Area {
    private long areaID;
    private String areaCode;
    private String areaName;
    private Timestamp createTime;

    public long getAreaID() {
        return areaID;
    }

    public void setAreaID(long areaID) {
        this.areaID = areaID;
    }

    public String getAreaCode() {
        return areaCode;
    }

    public void setAreaCode(String areaCode) {
        this.areaCode = areaCode;
    }

    public String getAreaName() {
        return areaName;
    }

    public void setAreaName(String areaName) {
        this.areaName = areaName;
    }

    public Timestamp getCreateTime() {
        return createTime;
    }

    public void setCreateTime(Timestamp createTime) {
        this.createTime = createTime;
    }
}
